+++ BestVehicleSeller/VehicleShop/Interfaces/IElectrical.cs
@@ -3,7 +3,7 @@
     public interface IElectrical
     {
         int BatteryCapacity { get; }    // in kWh
-        int MotorPower { get; }         // in kW
+        double MotorPower { get; }         // in kW
         int Range { get; }              // in km
     }
 }
